restructure pyspark.sql
